Unread/read messages work. Workingh on autofilling form when sending message.

+++ src/insertion/database.sql
@@ -23,7 +23,7 @@ CREATE TABLE messages (
 	messageID INT(11) PRIMARY KEY AUTO_INCREMENT,
     senderName VARCHAR (255) NOT NULL,
     recipientName VARCHAR (255) NOT NULL,
-    timeSent INT (11) NOT NULL,
+    timeSent BIGINT (11) NOT NULL,
     mSubject VARCHAR (255) NOT NULL,
     mContent TEXT NOT NULL,
 	mRead BIT NOT NULL DEFAULT 0,
